refactor(lab/13): move dom stuff into it's own module, add a web test
