+++ src/lab/13-game-of-life/src/lib.rs
@@ -2,52 +2,22 @@ extern crate js_sys;
 extern crate wasm_bindgen;
 extern crate web_sys;
 
+mod dom;
 mod life;
 mod util;
 
-use js_sys::Object;
-use life::{Cell, Universe};
+use dom::{canvas, ctx, raf};
+use life::Universe;
 use std::cell::RefCell;
 use std::rc::Rc;
 use util::set_panic_hook;
 use wasm_bindgen::prelude::*;
-use wasm_bindgen::JsCast;
-use web_sys::{window, CanvasRenderingContext2d, Document, Element, HtmlCanvasElement, Window};
 
 // When the `wee_alloc` feature is enabled, use `wee_alloc` as the global
 // allocator.
 #[cfg(feature = "wee_alloc")]
 #[global_allocator]
 static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
-
-fn win() -> Window {
-    window().expect("window to exist")
-}
-
-fn doc() -> Document {
-    win().document().expect("window.document to exist")
-}
-
-fn raf(f: &Closure<FnMut()>) {
-    win()
-        .request_animation_frame(f.as_ref().unchecked_ref())
-        .expect("window.requestAnimationFrame to exist");
-}
-
-fn canvas() -> Result<HtmlCanvasElement, Element> {
-    doc()
-        .query_selector("canvas")?
-        .unwrap()
-        .dyn_into::<HtmlCanvasElement>()
-}
-
-fn ctx(canvas: &HtmlCanvasElement) -> Result<CanvasRenderingContext2d, Object> {
-    canvas
-        .get_context("2d")
-        .unwrap()
-        .unwrap()
-        .dyn_into::<CanvasRenderingContext2d>()
-}
 
 #[wasm_bindgen]
 pub fn main() -> Result<(), JsValue> {
@@ -61,35 +31,13 @@ pub fn main() -> Result<(), JsValue> {
 
     let mut uni = Universe::new();
 
-    let size = 1;
-    let dead = JsValue::from_str("orange");
-    let live = JsValue::from_str("yellow");
-
     let f = Rc::new(RefCell::new(None));
     let g = f.clone();
 
     *g.borrow_mut() = Some(Closure::wrap(Box::new(move || {
         raf(f.borrow().as_ref().unwrap());
-
-        uni.tick();
-        for row in 0..uni.height {
-            for col in 0..uni.width {
-                let i = uni.get_index(row, col);
-
-                context.set_fill_style(if uni.cells[i] == Cell::Dead {
-                    &dead
-                } else {
-                    &live
-                });
-
-                context.fill_rect(
-                    (col * (size + 1) + 1).into(),
-                    (row * (size + 1) + 1).into(),
-                    size.into(),
-                    size.into(),
-                );
-            }
-        }
+        uni.update();
+        uni.render(&context);
     }) as Box<FnMut()>));
 
     raf(g.borrow().as_ref().unwrap());
